requirements.txt
argon2-cffi
cryptography
click 